requirements.txt
celery
redis
pymupdf
python-pptx
python-docx
httpx
pymongo
python-dotenv
requests
google-genai
json-repair
